Expect utf-8 bytes as input and make output match reference implementation
This fixes the issues with the value of multi-byte character codes being
partially ignored by the JS implementation.

The signature of all three variants of the hash function has been changed; now
they expect bytes instead of strings. This increases the flexibility of the
library as it can now operate on e.g. strings with arbitrary encodings, numbers
etc.

If the conversion from string to bytes was done internally, the library would
need to include at least a TextEncoder polyfill for utf-8 support and other
polyfills / hacks in order to have cross-browser support for other encodings.